before-final-demo some-changes

+++ SmartEventManagementSystem/Controllers/AttendeeController.cs
@@ -6,7 +6,6 @@ using Smart_Event_Management_System.CustomException;
 using Smart_Event_Management_System.Dto;
 using Smart_Event_Management_System.Models;
 using Smart_Event_Management_System.Service;
-using Smart_Event_Management_System.Validators;
 
 namespace Smart_Event_Management_System.Controllers;
 
@@ -35,6 +34,11 @@ public class AttendeeController : ControllerBase
     [HttpGet("{id}")]
     public async Task<IActionResult> GetAttendee(int id)
     {
+        if (id <= 0)
+        {
+            throw new InvalidIDException("Attendee Id must be greater than zero.");
+        }
+        
         var attendee = await _service.GetAttendeeByIdAsync(id);
         return Ok(attendee);
     }
@@ -66,6 +70,27 @@ public class AttendeeController : ControllerBase
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAttendee(int id, [FromBody] Attendee updatedAttendee)
     {
+        if (id <= 0)
+        {
+            throw new InvalidIDException("Attendee Id must be greater than zero.");
+        }
+        
+        var validationResult = await _attendeeValidator.ValidateAsync(updatedAttendee);
+
+        if (!validationResult.IsValid) {
+            var errors = validationResult.Errors
+                .Select(error => new ValidationFailure(error.PropertyName, error.ErrorMessage))
+                .ToList();
+
+            var validationErrorResponse = new ValidationErrorResponse()
+            {
+                Message = "Validation errors occurred.",
+                Errors = errors
+            };
+
+            return BadRequest(validationErrorResponse);
+        }
+        
         var result = await _service.UpdateAttendeeAsync(id, updatedAttendee);
 
         if (!result) return NotFound();
@@ -77,6 +102,11 @@ public class AttendeeController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAttendee(int id)
     {
+        if (id <= 0)
+        {
+            throw new InvalidIDException("Attendee Id must be greater than zero.");
+        }
+        
         await _service.DeleteAttendeeAsync(id);
         return NoContent();
     }
